+++ src/main/java/ru/dsoccer1980/domain/Genre.java
@@ -1,15 +1,15 @@
 package ru.dsoccer1980.domain;
 
 public class Genre {
-    private int id;
+    private long id;
     private String name;
 
-    public Genre(int id, String name) {
+    public Genre(long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
